# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Local Operator API",
    "description": "REST API interface for Local Operator agent",
    "version": "0.11.0"
  },
  "paths": {
...
    "/v1/chat": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Process chat request",
        "description": "Accepts a prompt and optional context/configuration, returns the model response and conversation history.",
        "operationId": "chat_endpoint_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              },
              "examples": {
                "example": {
                  "summary": "Example Request",
                  "value": {
                    "prompt": "Print 'Hello, world!'",
                    "hosting": "openrouter",
                    "model": "google/gemini-2.0-flash-001",
                    "context": [],
                    "options": {
                      "temperature": 0.2,
                      "top_p": 0.9
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CRUDResponse_ChatResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/agents/{agent_id}": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Process chat request using a specific agent",
        "description": "Accepts a prompt and optional context/configuration, retrieves the specified agent from the registry, applies it to the operator and executor, and returns the model response and conversation history.",
        "operationId": "chat_with_agent_v1_chat_agents__agent_id__post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID of the agent to use for the chat",
              "examples": [
                "agent123"
              ],
              "title": "Agent Id"
            },
            "description": "ID of the agent to use for the chat"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatRequest"
              },
              "examples": {
                "example": {
                  "summary": "Example Request with Agent",
                  "value": {
                    "prompt": "How do I implement a binary search in Python?",
                    "hosting": "openrouter",
                    "model": "google/gemini-2.0-flash-001",
                    "options": {
                      "temperature": 0.2,
                      "top_p": 0.9
                    },
                    "persist_conversation": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CRUDResponse_ChatResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/async": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Process chat request asynchronously",
        "description": "Accepts a prompt and optional context/configuration, starts an asynchronous job to process the request and returns a job ID.",
        "operationId": "chat_async_endpoint_v1_chat_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              },
              "examples": {
                "example": {
                  "summary": "Example Async Request",
                  "value": {
                    "prompt": "Print 'Hello, world!'",
                    "hosting": "openrouter",
                    "model": "google/gemini-2.0-flash-001",
                    "context": [],
                    "options": {
                      "temperature": 0.2,
                      "top_p": 0.9
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CRUDResponse_JobResultSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/agents/{agent_id}/async": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Process agent chat request asynchronously",
        "description": "Accepts a prompt and optional context/configuration, retrieves the specified agent from the registry, starts an asynchronous job to process the request and returns a job ID.",
        "operationId": "chat_with_agent_async_v1_chat_agents__agent_id__async_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID of the agent to use for the chat",
              "examples": [
                "agent123"
              ],
              "title": "Agent Id"
            },
            "description": "ID of the agent to use for the chat"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatRequest"
              },
              "examples": {
                "example": {
                  "summary": "Example Async Agent Request",
                  "value": {
                    "prompt": "How do I implement a binary search in Python?",
                    "hosting": "openrouter",
                    "model": "google/gemini-2.0-flash-001",
                    "options": {
                      "temperature": 0.2,
                      "top_p": 0.9
                    },
                    "persist_conversation": false,
                    "user_message_id": ""
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CRUDResponse_JobResultSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentChatRequest": {
        "properties": {
          "hosting": {
            "type": "string",
            "title": "Hosting"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "persist_conversation": {
            "type": "boolean",
            "title": "Persist Conversation",
            "default": false
          },
          "user_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Message Id"
          },
          "attachments": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "hosting",
          "model",
          "prompt"
        ],
        "title": "AgentChatRequest",
        "description": "Request body for chat generation endpoint.\n\nAttributes:\n    hosting: Name of the hosting service to use for generation\n    model: Name of the model to use for generation\n    prompt: The prompt to generate a response for\n    stream: Whether to stream the response token by token. Default: False\n    options: Optional generation parameters to override defaults\n    persist_conversation: Whether to persist the conversation history by\n    continuously updating the agent's conversation history with each new message.\n    Default: False\n    user_message_id: Optional ID of the user message to assign to the first user message\n        in the conversation.  This is used by the UI to prevent duplicate user\n        messages after the initial render.\n    attachments: Optional list of file paths (local or remote) to be used in the analysis.\n        These files are expected to be publicly accessible."
      },
...
      "CRUDResponse_ChatResponse_": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "CRUDResponse[ChatResponse]"
      },
...
      "CRUDResponse_JobResultSchema_": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobResultSchema"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "CRUDResponse[JobResultSchema]"
      },
...
      "ChatRequest": {
        "properties": {
          "hosting": {
            "type": "string",
            "title": "Hosting"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          },
          "context": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ConversationRecord"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "attachments": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "hosting",
          "model",
          "prompt"
        ],
        "title": "ChatRequest",
        "description": "Request body for chat generation endpoint.\n\nAttributes:\n    hosting: Name of the hosting service to use for generation\n    model: Name of the model to use for generation\n    prompt: The prompt to generate a response for\n    stream: Whether to stream the response token by token. Default: False\n    context: Optional list of previous messages for context\n    options: Optional generation parameters to override defaults\n    attachments: Optional list of file paths (local or remote) to be used in the analysis.\n        These files are expected to be publicly accessible."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}